refactor parser modules
